Refactor: Queue 관련 MapStruct 라이브러리 적용 코드삭제 , Dto 경로수정, 기타 코드 리팩터링
- MapStruct를 사용한 코드를삭제하고 .of() 메서드를 만들어 적용했습니다.
- Dto의 경로를 수정했습니다.
